==> Code/0.3_R_code-clinical_data_convert_to_csv.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

# Install R and R magic
!apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev
%load_ext rpy2.ipython

your_ds_project_path = '/content/drive/MyDrive/DS_project'

%%R
#GBM

# Set working directory to where your .RData file is stored

# Define the base folder path to your project directory in Google Drive
raw_data_path <- '/content/drive/MyDrive/DS_project/Data/1_raw_data'
clinical_data_path <- '/content/drive/MyDrive/DS_project/Data/0.clinical_rdata_to_csv'

# Load the RData file (e.g., patient clinical data)
load(paste0(raw_data_path, "/","TCGA-GBM_clinicalXML_patient.rdata"))
load(paste0(raw_data_path, "/","TCGA-GBM_clinical_indexed.rdata"))
load(paste0(raw_data_path, "/","TCGA-GBM_clinicalXML_admin.rdata"))
load(paste0(raw_data_path, "/","TCGA-GBM_clinicalXML_drug.rdata"))
load(paste0(raw_data_path, "/","TCGA-GBM_clinicalXML_followUp.rdata"))
load(paste0(raw_data_path, "/","TCGA-GBM_clinicalXML_newTumorEvent.rdata"))
load(paste0(raw_data_path, "/","TCGA-GBM_clinicalXML_radiation.rdata"))
load(paste0(raw_data_path, "/","TCGA-LGG_clinicalXML_stageEvent.rdata"))

# Export to CSV
write.csv(clinicalXML_patient ,paste0(clinical_data_path, "/","TCGA-GBM_clinicalXML_patient.csv"), row.names = FALSE)
write.csv(clinical_indexed,paste0(clinical_data_path, "/","TCGA-GBM_clinical_indexed.csv"), row.names = FALSE)
write.csv(clinicalXML_admin,paste0(clinical_data_path, "/","TCGA-GBM_clinicalXML_admin.csv"), row.names = FALSE)
write.csv(clinicalXML_drug,paste0(clinical_data_path, "/","TCGA-GBM_clinicalXML_drug.csv"), row.names = FALSE)
write.csv(clinicalXML_followUp, paste0(clinical_data_path, "/","TCGA-GBM_clinicalXML_followUp.csv"), row.names = FALSE)
write.csv(clinicalXML_newTumorEvent ,paste0(clinical_data_path, "/","TCGA-GBM_clinicalXML_newTumorEvent.csv"), row.names = FALSE)
write.csv(clinicalXML_radiation,paste0(clinical_data_path, "/","TCGA-GBM_clinicalXML_radiation.csv"), row.names = FALSE)
write.csv(clinicalXML_stageEvent,paste0(clinical_data_path, "/","TCGA-LGG_clinicalXML_stageEvent.csv"), row.names = FALSE)


%%R
#LGG

# Set working directory to where your .RData file is stored

# Define the base folder path to your project directory in Google Drive
raw_data_path <- '/content/drive/MyDrive/DS_project/Data/1_raw_data'
clinical_data_path <- '/content/drive/MyDrive/DS_project/Data/0.clinical_rdata_to_csv'


load(paste0(raw_data_path, "/","TCGA-LGG_clinicalXML_patient.rdata"))
load(paste0(raw_data_path, "/","TCGA-LGG_clinical_indexed.rdata"))
load(paste0(raw_data_path, "/","TCGA-LGG_clinicalXML_admin.rdata"))
load(paste0(raw_data_path, "/","TCGA-LGG_clinicalXML_drug.rdata"))
load(paste0(raw_data_path, "/","TCGA-LGG_clinicalXML_followUp.rdata"))
load(paste0(raw_data_path, "/","TCGA-LGG_clinicalXML_newTumorEvent.rdata"))
load(paste0(raw_data_path, "/","TCGA-LGG_clinicalXML_radiation.rdata"))
load(paste0(raw_data_path, "/","TCGA-LGG_clinicalXML_stageEvent.rdata"))


# Export to CSV
write.csv(clinicalXML_patient ,paste0(clinical_data_path, "/","TCGA-LGG_clinicalXML_patient.csv"), row.names = FALSE)
write.csv(clinical_indexed$clinical_indexed,paste0(clinical_data_path, "/","TCGA-LGG_clinical_indexed.csv"), row.names = FALSE)
write.csv(clinicalXML_admin,paste0(clinical_data_path, "/","TCGA-LGG_clinicalXML_admin.csv"), row.names = FALSE)
write.csv(clinicalXML_drug,paste0(clinical_data_path, "/","TCGA-LGG_clinicalXML_drug.csv"), row.names = FALSE)
write.csv(clinicalXML_followUp, paste0(clinical_data_path, "/","TCGA-LGG_clinicalXML_followUp.csv"), row.names = FALSE)
write.csv(clinicalXML_newTumorEvent ,paste0(clinical_data_path, "/","TCGA-LGG_clinicalXML_newTumorEvent.csv"), row.names = FALSE)
write.csv(clinicalXML_radiation,paste0(clinical_data_path, "/","TCGA-LGG_clinicalXML_radiation.csv"), row.names = FALSE)
write.csv(clinicalXML_stageEvent,paste0(clinical_data_path, "/","TCGA-LGG_clinicalXML_stageEvent.csv"), row.names = FALSE)
