==> Code/5.1_Multi-Omics_data_integration.ipynb <==
# prompt: import drive

from google.colab import drive
drive.mount('/content/drive')

!pip install mofapy2
!pip install anndata
!pip install mofax

from sklearn.preprocessing import StandardScaler

import mofax as mfx

import pandas as pd
import numpy as np
from mofapy2.run.entry_point import entry_point
import pickle

# Define the base folder path to your project directory in Google Drive
Ds_project_folder_path = '/content/drive/MyDrive/DS_project/'
data_spliting_meta_data_path = 'Data/5_Inter-cohort_preprocessing/5.1.data_spliting/'
splited_data_save_path = 'Data/5_Inter-cohort_preprocessing/5.2.data_cleaning(Feature_filtering)/'
batch_correction_save_path = 'Data/5_Inter-cohort_preprocessing/5.3.batch_correction/'
feature_selection_save_path = 'Data/5_Inter-cohort_preprocessing/5.4.feature_selection/'
data_integration_save_path = 'Data/5_Inter-cohort_preprocessing/5.5Multi-omics_integration/'

feature_selection_file_path = Ds_project_folder_path + feature_selection_save_path
split_file_path = Ds_project_folder_path + splited_data_save_path
data_integration_file_path = Ds_project_folder_path + data_integration_save_path

version = '_top_1000'

### MOFA Train

GBM_SNV_train = pd.read_csv(f'{split_file_path}GBM_SNV_train.csv', index_col=0)
GBM_SNV_val= pd.read_csv(f'{split_file_path}GBM_SNV_val.csv', index_col=0)
GBM_SNV_external_test= pd.read_csv(f'{split_file_path}GBM_SNV_external_test.csv', index_col=0)
GBM_SNV_internal_test= pd.read_csv(f'{split_file_path}GBM_SNV_internal_test.csv', index_col=0)

LGG_SNV_train= pd.read_csv(f'{split_file_path}LGG_SNV_train.csv', index_col=0)
LGG_SNV_val= pd.read_csv(f'{split_file_path}LGG_SNV_val.csv', index_col=0)
LGG_SNV_external_test= pd.read_csv(f'{split_file_path}LGG_SNV_external_test.csv', index_col=0)
LGG_SNV_internal_test= pd.read_csv(f'{split_file_path}LGG_SNV_internal_test.csv', index_col=0)

#CNV
GBM_CNV_train= pd.read_csv(f'{feature_selection_file_path}GBM_CNV_train_feature_selected{version}.csv', index_col=0)
LGG_CNV_train= pd.read_csv(f'{feature_selection_file_path}LGG_CNV_train_feature_selected{version}.csv', index_col=0)

#Methlyation
GBM_methylation_train = pd.read_csv(f'{feature_selection_file_path}GBM_methylation_train_feature_selected{version}.csv', index_col=0)
LGG_methylation_train = pd.read_csv(f'{feature_selection_file_path}LGG_methylation_train_feature_selected{version}.csv', index_col=0)

#RNA_seq
GBM_rna_seq_train= pd.read_csv(f'{feature_selection_file_path}GBM_rna_seq_train_feature_selected{version}.csv', index_col=0)
LGG_rna_seq_train= pd.read_csv(f'{feature_selection_file_path}LGG_rna_seq_train_feature_selected{version}.csv', index_col=0)

#Protein
#GBM_protein_train= pd.read_csv(f'{split_file_path}GBM_protein_train{version}.csv', index_col=0)
#LGG_protein_train= pd.read_csv(f'{split_file_path}LGG_protein_train{version}.csv', index_col=0)
GBM_protein_train= pd.read_csv(f'{feature_selection_file_path}GBM_protein_train_feature_selected{version}.csv', index_col=0)
LGG_protein_train= pd.read_csv(f'{feature_selection_file_path}LGG_protein_train_feature_selected{version}.csv', index_col=0)

#SNV
#GBM_SNV_train= pd.read_csv(f'{split_file_path}GBM_SNV_train{version}.csv', index_col=0)
#LGG_SNV_train= pd.read_csv(f'{split_file_path}LGG_SNV_train{version}.csv', index_col=0)
#GBM_SNV_train= pd.read_csv(f'{feature_selection_file_path}GBM_SNV_train_feature_selected{version}.csv', index_col=0)
#LGG_SNV_train= pd.read_csv(f'{feature_selection_file_path}LGG_SNV_train_feature_selected{version}.csv', index_col=0)

#CNV
GBM_CNV_val= pd.read_csv(f'{feature_selection_file_path}GBM_CNV_val_feature_selected{version}.csv', index_col=0)
GBM_CNV_external_test= pd.read_csv(f'{feature_selection_file_path}GBM_CNV_external_test_feature_selected{version}.csv', index_col=0)
GBM_CNV_internal_test= pd.read_csv(f'{feature_selection_file_path}GBM_CNV_internal_test_feature_selected{version}.csv', index_col=0)

LGG_CNV_val= pd.read_csv(f'{feature_selection_file_path}LGG_CNV_val_feature_selected{version}.csv', index_col=0)
LGG_CNV_external_test= pd.read_csv(f'{feature_selection_file_path}LGG_CNV_external_test_feature_selected{version}.csv', index_col=0)
LGG_CNV_internal_test= pd.read_csv(f'{feature_selection_file_path}LGG_CNV_internal_test_feature_selected{version}.csv', index_col=0)

#Methlyation
GBM_methylation_val= pd.read_csv(f'{feature_selection_file_path}GBM_methylation_val_feature_selected{version}.csv', index_col=0)
GBM_methylation_external_test= pd.read_csv(f'{feature_selection_file_path}GBM_methylation_external_test_feature_selected{version}.csv', index_col=0)
GBM_methylation_internal_test= pd.read_csv(f'{feature_selection_file_path}GBM_methylation_internal_test_feature_selected{version}.csv', index_col=0)

LGG_methylation_val= pd.read_csv(f'{feature_selection_file_path}LGG_methylation_val_feature_selected{version}.csv', index_col=0)
LGG_methylation_external_test= pd.read_csv(f'{feature_selection_file_path}LGG_methylation_external_test_feature_selected{version}.csv', index_col=0)
LGG_methylation_internal_test= pd.read_csv(f'{feature_selection_file_path}LGG_methylation_internal_test_feature_selected{version}.csv', index_col=0)

#RNA_seq
GBM_rna_seq_val= pd.read_csv(f'{feature_selection_file_path}GBM_rna_seq_val_feature_selected{version}.csv', index_col=0)
GBM_rna_seq_external_test= pd.read_csv(f'{feature_selection_file_path}GBM_rna_seq_external_test_feature_selected{version}.csv', index_col=0)
GBM_rna_seq_internal_test= pd.read_csv(f'{feature_selection_file_path}GBM_rna_seq_internal_test_feature_selected{version}.csv', index_col=0)

LGG_rna_seq_val= pd.read_csv(f'{feature_selection_file_path}LGG_rna_seq_val_feature_selected{version}.csv', index_col=0)
LGG_rna_seq_external_test= pd.read_csv(f'{feature_selection_file_path}LGG_rna_seq_external_test_feature_selected{version}.csv', index_col=0)
LGG_rna_seq_internal_test= pd.read_csv(f'{feature_selection_file_path}LGG_rna_seq_internal_test_feature_selected{version}.csv', index_col=0)

#Protein
#GBM_protein_val= pd.read_csv(f'{split_file_path}GBM_protein_val.csv', index_col=0)
#GBM_protein_external_test= pd.read_csv(f'{split_file_path}GBM_protein_external_test.csv', index_col=0)
#GBM_protein_internal_test= pd.read_csv(f'{split_file_path}GBM_protein_internal_test.csv', index_col=0)

#LGG_protein_val= pd.read_csv(f'{split_file_path}LGG_protein_val.csv', index_col=0)
#LGG_protein_external_test= pd.read_csv(f'{split_file_path}LGG_protein_external_test.csv', index_col=0)
#LGG_protein_internal_test= pd.read_csv(f'{split_file_path}LGG_protein_internal_test.csv', index_col=0)

GBM_protein_val= pd.read_csv(f'{feature_selection_file_path}GBM_protein_val_feature_selected{version}.csv', index_col=0)
GBM_protein_external_test= pd.read_csv(f'{feature_selection_file_path}GBM_protein_external_test_feature_selected{version}.csv', index_col=0)
GBM_protein_internal_test= pd.read_csv(f'{feature_selection_file_path}GBM_protein_internal_test_feature_selected{version}.csv', index_col=0)

LGG_protein_val= pd.read_csv(f'{feature_selection_file_path}LGG_protein_val_feature_selected{version}.csv', index_col=0)
LGG_protein_external_test= pd.read_csv(f'{feature_selection_file_path}LGG_protein_external_test_feature_selected{version}.csv', index_col=0)
LGG_protein_internal_test= pd.read_csv(f'{feature_selection_file_path}LGG_protein_internal_test_feature_selected{version}.csv', index_col=0)

#SNV
#GBM_SNV_val= pd.read_csv(f'{split_file_path}GBM_SNV_val.csv', index_col=0)
#GBM_SNV_external_test= pd.read_csv(f'{split_file_path}GBM_SNV_external_test.csv', index_col=0)
#GBM_SNV_internal_test= pd.read_csv(f'{split_file_path}GBM_SNV_internal_test.csv', index_col=0)

#LGG_SNV_val= pd.read_csv(f'{split_file_path}LGG_SNV_val.csv', index_col=0)
#LGG_SNV_external_test= pd.read_csv(f'{split_file_path}LGG_SNV_external_test.csv', index_col=0)
#LGG_SNV_internal_test= pd.read_csv(f'{split_file_path}LGG_SNV_internal_test.csv', index_col=0)

#GBM_SNV_val= pd.read_csv(f'{feature_selection_file_path}GBM_SNV_val_feature_selected{version}.csv', index_col=0)
#GBM_SNV_external_test= pd.read_csv(f'{feature_selection_file_path}GBM_SNV_external_test_feature_selected{version}.csv', index_col=0)
#GBM_SNV_internal_test= pd.read_csv(f'{feature_selection_file_path}GBM_SNV_internal_test_feature_selected{version}.csv', index_col=0)

#LGG_SNV_val= pd.read_csv(f'{feature_selection_file_path}LGG_SNV_val_feature_selected{version}.csv', index_col=0)
#LGG_SNV_external_test= pd.read_csv(f'{feature_selection_file_path}LGG_SNV_external_test_feature_selected{version}.csv', index_col=0)
#LGG_SNV_internal_test= pd.read_csv(f'{feature_selection_file_path}LGG_SNV_internal_test_feature_selected{version}.csv', index_col=0)

version = "mixed"

CNV_val = pd.concat([GBM_CNV_val, LGG_CNV_val], axis=1)
CNV_external_test = pd.concat([GBM_CNV_external_test, LGG_CNV_external_test], axis=1)
CNV_internal_test = pd.concat([GBM_CNV_internal_test, LGG_CNV_internal_test], axis=1)
CNV_train = pd.concat([GBM_CNV_train, LGG_CNV_train], axis=1)

SNV_val = pd.concat([GBM_SNV_val, LGG_SNV_val], axis=1)
SNV_external_test = pd.concat([GBM_SNV_external_test, LGG_SNV_external_test], axis=1)
SNV_internal_test = pd.concat([GBM_SNV_internal_test, LGG_SNV_internal_test], axis=1)
SNV_train = pd.concat([GBM_SNV_train, LGG_SNV_train], axis=1)

methyl_val = pd.concat([GBM_methylation_val, LGG_methylation_val], axis=1)
methyl_external_test = pd.concat([GBM_methylation_external_test, LGG_methylation_external_test], axis=1)
methyl_internal_test = pd.concat([GBM_methylation_internal_test, LGG_methylation_internal_test], axis=1)
methyl_train = pd.concat([GBM_methylation_train, LGG_methylation_train], axis=1)

rna_seq_val = pd.concat([GBM_rna_seq_val, LGG_rna_seq_val], axis=1)
rna_seq_external_test = pd.concat([GBM_rna_seq_external_test, LGG_rna_seq_external_test], axis=1)
rna_seq_internal_test = pd.concat([GBM_rna_seq_internal_test, LGG_rna_seq_internal_test], axis=1)
rna_seq_train = pd.concat([GBM_rna_seq_train, LGG_rna_seq_train], axis=1)

protein_val = pd.concat([GBM_protein_val, LGG_protein_val], axis=1)
protein_external_test = pd.concat([GBM_protein_external_test, LGG_protein_external_test], axis=1)
protein_internal_test = pd.concat([GBM_protein_internal_test, LGG_protein_internal_test], axis=1)
protein_train = pd.concat([GBM_protein_train, LGG_protein_train], axis=1)

GBM_sample_sets = [
    GBM_CNV_train.columns, GBM_CNV_val.columns, GBM_CNV_external_test.columns, GBM_CNV_internal_test.columns,
    GBM_methylation_train.columns, GBM_methylation_val.columns, GBM_methylation_external_test.columns, GBM_methylation_internal_test.columns,
    GBM_rna_seq_train.columns, GBM_rna_seq_val.columns, GBM_rna_seq_external_test.columns, GBM_rna_seq_internal_test.columns,
    GBM_protein_train.columns, GBM_protein_val.columns, GBM_protein_external_test.columns, GBM_protein_internal_test.columns,
    GBM_SNV_train.columns, GBM_SNV_val.columns, GBM_SNV_external_test.columns, GBM_SNV_internal_test.columns
]

GBM_all_samples = sorted(set().union(*GBM_sample_sets))

LGG_sample_sets = [
    LGG_CNV_train.columns, LGG_CNV_val.columns, LGG_CNV_external_test.columns, LGG_CNV_internal_test.columns,
    LGG_methylation_train.columns, LGG_methylation_val.columns, LGG_methylation_external_test.columns, LGG_methylation_internal_test.columns,
    LGG_rna_seq_train.columns, LGG_rna_seq_val.columns, LGG_rna_seq_external_test.columns, LGG_rna_seq_internal_test.columns,
    LGG_protein_train.columns, LGG_protein_val.columns, LGG_protein_external_test.columns, LGG_protein_internal_test.columns,
    LGG_SNV_train.columns, LGG_SNV_val.columns, LGG_SNV_external_test.columns, LGG_SNV_internal_test.columns
]

LGG_all_samples = sorted(set().union(*LGG_sample_sets))

print("Total # of GBM samples:", len(GBM_all_samples))
print("Total # of LGG samples:", len(LGG_all_samples))

all_train_sample = sorted(set.union(set(CNV_train.columns),set(methyl_train.columns),set(rna_seq_train.columns),
                           set(protein_train.columns),set(SNV_train.columns)))
all_val_samples = sorted(set.union(set(CNV_val.columns),set(methyl_val.columns),set(rna_seq_val.columns),
                           set(protein_val.columns),set(SNV_val.columns)))
all_external_test_samples = sorted(set.union(set(CNV_external_test.columns),set(methyl_external_test.columns),set(rna_seq_external_test.columns),
                           set(protein_external_test.columns),set(SNV_external_test.columns)))
all_internal_test_samples = sorted(set.union(set(CNV_internal_test.columns),set(methyl_internal_test.columns),set(rna_seq_internal_test.columns),
                           set(protein_internal_test.columns),set(SNV_internal_test.columns)))

print("Total # of train samples:", len(all_train_sample))
print("Total # of val samples:", len(all_val_samples))
print("Total # of external test samples:", len(all_external_test_samples))
print("Total # of internal test samples:", len(all_internal_test_samples))

def align_omics_matrix(omics_df, sample_list):
    return omics_df.reindex(columns=sample_list, fill_value=np.nan)

data_dict_val = {
    "CNV": align_omics_matrix(CNV_val, all_val_samples),
    "Methylation": align_omics_matrix(methyl_val, all_val_samples),
    "RNAseq": align_omics_matrix(rna_seq_val, all_val_samples),
    "Protein": align_omics_matrix(protein_val, all_val_samples),
    "SNV": align_omics_matrix(SNV_val, all_val_samples)
}

data_dict_external_test = {
    "CNV": align_omics_matrix(CNV_external_test, all_external_test_samples),
    "Methylation": align_omics_matrix(methyl_external_test, all_external_test_samples),
    "RNAseq": align_omics_matrix(rna_seq_external_test, all_external_test_samples),
    "Protein": align_omics_matrix(protein_external_test, all_external_test_samples),
    "SNV": align_omics_matrix(SNV_external_test, all_external_test_samples)
}

data_dict_internal_test = {
    "CNV": align_omics_matrix(CNV_internal_test, all_internal_test_samples),
    "Methylation": align_omics_matrix(methyl_internal_test, all_internal_test_samples),
    "RNAseq": align_omics_matrix(rna_seq_internal_test, all_internal_test_samples),
    "Protein": align_omics_matrix(protein_internal_test, all_internal_test_samples),
    "SNV": align_omics_matrix(SNV_internal_test, all_internal_test_samples)
}

data_dict_train = {
    "CNV": align_omics_matrix(CNV_train, all_train_sample),
    "Methylation": align_omics_matrix(methyl_train, all_train_sample),
    "RNAseq": align_omics_matrix(rna_seq_train, all_train_sample),
    "Protein": align_omics_matrix(protein_train, all_train_sample),
    "SNV": align_omics_matrix(SNV_train, all_train_sample)
}

def zscore_normalize_df(df):
    means = df.mean(axis=1)
    stds = df.std(axis=1)
    stds_replaced = stds.replace(0, 1)
    return (df.sub(means, axis=0)).div(stds_replaced, axis=0)

def zscore_normalize_data_dict(data_dict, exclude_layers=[]):
    normalized_dict = {}
    for omic_type, df in data_dict.items():
        if omic_type in exclude_layers:
            normalized_dict[omic_type] = df.copy()
        elif df.isnull().all().all():
            normalized_dict[omic_type] = df.copy()
        else:
            normalized_dict[omic_type] = zscore_normalize_df(df)
    return normalized_dict

znor_train_dic = zscore_normalize_data_dict(data_dict_train)
znor_val_dic = zscore_normalize_data_dict(data_dict_val)
znor_external_test_dic = zscore_normalize_data_dict(data_dict_external_test)
znor_internal_test_dic = zscore_normalize_data_dict(data_dict_internal_test)

def prepare_mofa_input(data_dict):
    data_list = []
    for view_name, df in data_dict.items():
        df_long = df.stack().reset_index()
        df_long.columns = ['feature', 'sample', 'value']
        df_long['view'] = view_name
        data_list.append(df_long)
    return pd.concat(data_list, ignore_index=True)

mofa_train_df = prepare_mofa_input(znor_train_dic)

## MOFA Training

def latent_vector_val_test_Data(ori_data, mofa_train):
  Z = [0]
  for k, v in mofa_train.items():
    filled_data = ori_data[k].T.fillna(ori_data[k].T.mean()).T.to_numpy()
    Z.append(v.T @ filled_data)
  return Z

#factors = [5, 10, 15, 20 ,25, 26,27,28,29,30,31,32,33,34,35]
# factors = [26,27,28,29,30,31,32,33,34,35]
#factors = [5, 10, 15, 20 ,25, 26,27,28,29,30,31,32,33,34,35]
factors = [30]

for factor in factors:
  ent = entry_point()
  ent.set_data_df(mofa_train_df)
  ent.set_data_options(scale_views=True, use_float32=True)

  ent.set_data_options(scale_views=True, use_float32=True)
  ent.set_model_options(factors=factor)
  ent.set_train_options(iter=25, convergence_mode='medium', seed = 822578 )
  #ent.set_train_options(iter=25, convergence_mode='medium')

  ent.build()
  ent.run()

  ent.save(f"{data_integration_file_path}trained_model_factor_{str(factor)}{version}.hdf5")
  ex_parameter = {layer_name:e for layer_name, e in zip(['CNV', 'Methylation','Protein','RNAseq', 'SNV'],ent.model.nodes['W'].getExpectation())}

  val_Z = latent_vector_val_test_Data(data_dict_val,ex_parameter)
  external_Z = latent_vector_val_test_Data(data_dict_external_test,ex_parameter)
  internal_test_Z = latent_vector_val_test_Data(data_dict_internal_test,ex_parameter)


  val_latent_vector = np.zeros((factor, len(data_dict_val['CNV'].columns)))
  external_latent_vector = np.zeros((factor, len(data_dict_external_test['CNV'].columns)))
  internal_test_latent_vector = np.zeros((factor, len(data_dict_internal_test['CNV'].columns)))

  for omic_val, omics_external, omics_internal in zip(val_Z, external_Z, internal_test_Z):
    val_latent_vector += omic_val
    external_latent_vector += omics_external
    internal_test_latent_vector += omics_internal

  val_latent_vector /= len(val_Z)
  external_latent_vector /= len(external_Z)
  internal_test_latent_vector /= len(internal_test_Z)

  train_latent_output = []
  val_latent_output = []
  external_test_latent_output = []
  internal_test_latent_output = []
  for sample, vector in zip(data_dict_train['CNV'].columns , ent.model.nodes['Z'].getExpectation()):
    if sample in GBM_all_samples:
      dd = {'sample_id':sample, 'latent_vector':vector, 'label': 'GBM'}
    elif sample in LGG_all_samples:
      dd = {'sample_id':sample, 'latent_vector':vector, 'label': 'LGG'}
    train_latent_output.append(dd)

  for sample, vector in zip(data_dict_val['CNV'].columns , val_latent_vector.T):
    if sample in GBM_all_samples:
      dd = {'sample_id':sample, 'latent_vector':vector, 'label': 'GBM'}
    elif sample in LGG_all_samples:
      dd = {'sample_id':sample, 'latent_vector':vector, 'label': 'LGG'}
    val_latent_output.append(dd)

  for sample, vector in zip(data_dict_external_test['CNV'].columns , external_latent_vector.T):
    if sample in GBM_all_samples:
      dd = {'sample_id':sample, 'latent_vector':vector, 'label': 'GBM'}
    elif sample in LGG_all_samples:
      dd = {'sample_id':sample, 'latent_vector':vector, 'label': 'LGG'}
    external_test_latent_output.append(dd)

  for sample, vector in zip(data_dict_internal_test['CNV'].columns , internal_test_latent_vector.T):
    if sample in GBM_all_samples:
      dd = {'sample_id':sample, 'latent_vector':vector, 'label': 'GBM'}
    elif sample in LGG_all_samples:
      dd = {'sample_id':sample, 'latent_vector':vector, 'label': 'LGG'}
    internal_test_latent_output.append(dd)

  with open(data_integration_file_path + 'Z_output/' + str(factor) + version + '_train_latent_output.pickle', 'wb') as f:
    pickle.dump(train_latent_output, f, protocol=pickle.HIGHEST_PROTOCOL)
  with open(data_integration_file_path + 'Z_output/' +str(factor) + version + '_val_latent_output.pickle', 'wb') as f:
      pickle.dump(val_latent_output, f, protocol=pickle.HIGHEST_PROTOCOL)
  with open(data_integration_file_path +'Z_output/' +str(factor) + version + '_external_test_latent_output.pickle', 'wb') as f:
      pickle.dump(external_test_latent_output, f, protocol=pickle.HIGHEST_PROTOCOL)
  with open(data_integration_file_path +'Z_output/' +str(factor) + version + '_internal_test_latent_output.pickle', 'wb') as f:
      pickle.dump(internal_test_latent_output, f, protocol=pickle.HIGHEST_PROTOCOL)
